chore: rename cloud-cognitive to ibm-products
