+++ config/jest-config-ibm-cloud-cognitive/setup/setupFiles.js
@@ -5,8 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-// Avoid side effects from `Canary` in `cloud-cognitive/src/settings`.
-import settings from '../../../packages/cloud-cognitive/src/global/js/package-settings';
+// Avoid side effects from `Canary` in `ibm-products/src/settings`.
+import settings from '../../../packages/ibm-products/src/global/js/package-settings';
 
 const { _silenceWarnings, setAllComponents } = settings;
