+++ packages/cloud-cognitive/package.json
@@ -8,7 +8,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/carbon-design-system/ibm-products.git",
-    "directory": "packages/cloud-cognitive"
+    "directory": "packages/ibm-products"
   },
   "bugs": "https://github.com/carbon-design-system/ibm-products/issues",
   "sideEffects": [
