+++ packages/component-usage-report/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@carbon/component-usage-report",
   "version": "0.2.1",
-  "description": "Creates a @carbon/cloud-cognitive component usage report.",
+  "description": "Creates a @carbon/ibm-products component usage report.",
   "bin": "bin/component-usage-report",
   "publishConfig": {
     "access": "public"
