+++ packages/core/.storybook/carbon.scss
@@ -7,5 +7,5 @@
 
 // a build of carbon css with our required feature flags set
 
-@import '../../cloud-cognitive/src/global/styles/project-settings';
+@import '../../ibm-products/src/global/styles/project-settings';
 @import 'carbon-components/scss/globals/scss/styles';
