+++ packages/core/.storybook/index.scss
@@ -9,7 +9,7 @@
 // Carbon styles at the end -- this is to obtain a "worst case" for our own CSS,
 // to ensure we are resilient against different CSS loading orders and our
 // styles have the specificity necessary to override Carbon styles when needed.
-@import '../../cloud-cognitive/src/index-without-carbon';
+@import '../../ibm-products/src/index-without-carbon';
 
 // Load all Carbon styles now
 @import '../css/carbon';
