+++ packages/core/src/component-playground/_storybook-styles.scss
@@ -5,7 +5,7 @@
 // LICENSE file in the root directory of this source tree.
 //
 
-@import '../../../cloud-cognitive/src/global/styles/project-settings';
+@import '../../../ibm-products/src/global/styles/project-settings';
 @import '@carbon/layout/scss/layout';
 
 .component-playground .#{$pkg-prefix}--side-panel__overlay {
